+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## v3.3.0
 
+- [Report diagnostics only on certain files with `reportFiles` option](https://github.com/TypeStrong/ts-loader/pull/701) - thanks @freeman!
 - [Added support for the new watch api of TypeScript compiler.](https://github.com/TypeStrong/ts-loader/pull/685)  nb This feature has been placed behind a new `experimentalWatchApi` option until it has been thoroughly tested. All being well it is likely to become the default behaviour for ts-loader in future - thanks @sheetalkamat!
 
 ## v3.2.0